Move to HA configuration for HSM
We enable HA-only mode for the HSM and use the active partition name for
the group label. This way, CredHub still logs into "the partition" but
"the partition" is really the virtual HA partition/slot, not the
physical slot.

HSM partitions/slots *seem* on some levels to be synoymous. However, an
HA partition gets represented as n + 1 slots: 1 "physical" slot per HSM
in the n-node cluster, where the name is the name of the partition on
the HSM; and 1 "virtual"/HA slot that only exists on the client machine.
The virtual slot is defined in /etc/Chrystoki.conf and goes by the
name/label of the HA group.

HA-only mode will only show the "virtual" slot to the CredHub app.

We also explicitly for an existing HA group with the expected name when
determining whether we need to add a new HA group.

If the HSMs in the cluster have changed, there should be a new
encryption.conf and therefore the prior HA group will be wiped out
(since HA config only exists insofar as it's in /etc/Chrystoki.conf).

If the HSMs haven't changed - including the partition name/password - we
should be able to simply keep the existing HA group.

[#136539551] CredHub should work appropriately with HA HSM

Signed-off-by: Sebastian Wittenkamp <swittenkamp@pivotal.io>