Only check encryption password length on active key
Previously our password length check was not validating the proper
field, and as such some CredHub servers were deployed with less than 20
character encryption passwords. In a recent release, we fixed the 20
character enforcement bug and some users were unable to upgrade. This
commit allows for a path to move from the affected versions to the
latest release by allowing an operator to make their less than 20
character key an inactive key and a new active key as a 20+ character
passphrase.

[#166990551]

Signed-off-by: Mark DeLillo <mdelillo@pivotal.io>