move trust store to env var
when passing the trust store password as an argument to the java command
it could be seen when running the `ps` command.

to prevent this password from being visible we moved the trust store
arguments to an environment variable that java will use.

we reorganized the file now that the rendering of the java command no
longer need to worry about the truststore variables.

[#173974580](https://www.pivotaltracker.com/story/show/173974580)

Signed-off-by: Peter Chen <pchen@pivotal.io>