feat: augment TLS compatibility with backends
- Before this commit, CredHub specifies that "secp384r1"
elliptic curve be used for TLS connections. We found no clear
justifications for this specification (the commit and story
associated with this specification, #154184086, did not explain
why).
- We decide to remove this specification, and instead,
rely on Java's default elliptic curve name groups,
which is an ordered list of elliptic curve groups deemed
acceptable by Java.
- We make this commit in preparation of bumping Java 8
patch to "bellsoft jre8u372" which might result in newer
cipher suites to be used in TLS connections. These new
cipher suites might require the client and server to use
a matching elliptic curve name group. This is a problem when
CredHub restricts its elliptic curve named groups to only
"secp384r1", which the server might not support. (This issue
manifested itself when we bumped the Java patch, CredHub
ran into TLS handshake failure with postgres-release with
the error "could not accept SSL connection: no suitable key share")

[#183852496]
[Related: https://github.com/pivotal/credhub-release/issues/74]